requirements.txt
Django==1.11.3
python-decouple>=3.0
django-compressor>=2.1
django-compressor-toolkit>=0.6.0
django-npm>=1.0.0
django-nested-admin>=3.0.19
django-tastypie>=0.14.0
django-js-reverse>=0.7.3
requests>=2.18.3